support char16_t in C++ if available, and wchar_t on windows

+++ library/include/pmath-language/tokens.h
@@ -91,32 +91,32 @@ enum {
   PMATH_PREC_PRIM =     1000
 };
 
-#define PMATH_CHAR_PLUSMINUS                 0x00B1 ///< The "+/-" character
-#define PMATH_CHAR_TIMES                     0x00D7 ///< The multiplication character
-#define PMATH_CHAR_INVISIBLECALL             0x2061 ///< The Function application character
-#define PMATH_CHAR_INVISIBLETIMES            0x2062 ///< The invisible "*" character
-#define PMATH_CHAR_INVISIBLECOMMA            0x2063 ///< The invisible "," character
-#define PMATH_CHAR_INVISIBLEPLUS             0x2064 ///< The invisible "+" character
-#define PMATH_CHAR_VECTOR                    0x21C0 ///< The arrow above names to indicate a vector
-#define PMATH_CHAR_RULE                      0x2192 ///< The "->" operator
-#define PMATH_CHAR_RULEDELAYED               0x29F4 ///< The ":>" operator
-#define PMATH_CHAR_ASSIGN                    0x2254 ///< The ":=" operator
-#define PMATH_CHAR_ASSIGNDELAYED             0x2A74 ///< The "::=" operator
-#define PMATH_CHAR_INTEGRAL_D                0x2146 ///< The integral "d"
-#define PMATH_CHAR_PIECEWISE                 0xF361 ///< The left curly bracket for cases
-#define PMATH_CHAR_LEFTINVISIBLEBRACKET      0xF362 ///< TeX's "\left."
-#define PMATH_CHAR_RIGHTINVISIBLEBRACKET     0xF363 ///< TeX's "\right."
-#define PMATH_CHAR_LEFTBRACKETINGBAR         0xF603 ///< TeX's "\left|"
-#define PMATH_CHAR_RIGHTBRACKETINGBAR        0xF604 ///< TeX's "\right|"
-#define PMATH_CHAR_LEFTDOUBLEBRACKETINGBAR   0xF605 ///< TeX's "\left\|"
-#define PMATH_CHAR_RIGHTDOUBLEBRACKETINGBAR  0xF606 ///< TeX's "\right\|"
-#define PMATH_CHAR_ALIASDELIMITER            0xF764 ///< The character inserted by Richmath with ESCAPE or CAPSLOCK
-#define PMATH_CHAR_ALIASINDICATOR            0xF768 ///< A character that looks like PMATH_CHAR_ALIASDELIMITER but has no effect
-#define PMATH_CHAR_LEFT_BOX                  0xFFF9 ///< Start of box code inside a string.
-#define PMATH_CHAR_RIGHT_BOX                 0xFFFB ///< End of box code inside a string.
-#define PMATH_CHAR_BOX                       0xFDD0 ///< Represents a box.
-#define PMATH_CHAR_PLACEHOLDER               0xFFFD ///< The Placeholder character. In richmath, type CAPSLOCK pl CAPSLOCK to insert it.
-#define PMATH_CHAR_SELECTIONPLACEHOLDER      0xF527 ///< The selection placeholder character. In richmath, type CAPSLOCK spl CAPSLOCK to insert it.
+#define PMATH_CHAR_PLUSMINUS                 ((uint16_t) 0x00B1 ) ///< The "+/-" character
+#define PMATH_CHAR_TIMES                     ((uint16_t) 0x00D7 ) ///< The multiplication character
+#define PMATH_CHAR_INVISIBLECALL             ((uint16_t) 0x2061 ) ///< The Function application character
+#define PMATH_CHAR_INVISIBLETIMES            ((uint16_t) 0x2062 ) ///< The invisible "*" character
+#define PMATH_CHAR_INVISIBLECOMMA            ((uint16_t) 0x2063 ) ///< The invisible "," character
+#define PMATH_CHAR_INVISIBLEPLUS             ((uint16_t) 0x2064 ) ///< The invisible "+" character
+#define PMATH_CHAR_VECTOR                    ((uint16_t) 0x21C0 ) ///< The arrow above names to indicate a vector
+#define PMATH_CHAR_RULE                      ((uint16_t) 0x2192 ) ///< The "->" operator
+#define PMATH_CHAR_RULEDELAYED               ((uint16_t) 0x29F4 ) ///< The ":>" operator
+#define PMATH_CHAR_ASSIGN                    ((uint16_t) 0x2254 ) ///< The ":=" operator
+#define PMATH_CHAR_ASSIGNDELAYED             ((uint16_t) 0x2A74 ) ///< The "::=" operator
+#define PMATH_CHAR_INTEGRAL_D                ((uint16_t) 0x2146 ) ///< The integral "d"
+#define PMATH_CHAR_PIECEWISE                 ((uint16_t) 0xF361 ) ///< The left curly bracket for cases
+#define PMATH_CHAR_LEFTINVISIBLEBRACKET      ((uint16_t) 0xF362 ) ///< TeX's "\left."
+#define PMATH_CHAR_RIGHTINVISIBLEBRACKET     ((uint16_t) 0xF363 ) ///< TeX's "\right."
+#define PMATH_CHAR_LEFTBRACKETINGBAR         ((uint16_t) 0xF603 ) ///< TeX's "\left|"
+#define PMATH_CHAR_RIGHTBRACKETINGBAR        ((uint16_t) 0xF604 ) ///< TeX's "\right|"
+#define PMATH_CHAR_LEFTDOUBLEBRACKETINGBAR   ((uint16_t) 0xF605 ) ///< TeX's "\left\|"
+#define PMATH_CHAR_RIGHTDOUBLEBRACKETINGBAR  ((uint16_t) 0xF606 ) ///< TeX's "\right\|"
+#define PMATH_CHAR_ALIASDELIMITER            ((uint16_t) 0xF764 ) ///< The character inserted by Richmath with ESCAPE or CAPSLOCK
+#define PMATH_CHAR_ALIASINDICATOR            ((uint16_t) 0xF768 ) ///< A character that looks like PMATH_CHAR_ALIASDELIMITER but has no effect
+#define PMATH_CHAR_LEFT_BOX                  ((uint16_t) 0xFFF9 ) ///< Start of box code inside a string.
+#define PMATH_CHAR_RIGHT_BOX                 ((uint16_t) 0xFFFB ) ///< End of box code inside a string.
+#define PMATH_CHAR_BOX                       ((uint16_t) 0xFDD0 ) ///< Represents a box.
+#define PMATH_CHAR_PLACEHOLDER               ((uint16_t) 0xFFFD ) ///< The Placeholder character. In richmath, type CAPSLOCK pl CAPSLOCK to insert it.
+#define PMATH_CHAR_SELECTIONPLACEHOLDER      ((uint16_t) 0xF527 ) ///< The selection placeholder character. In richmath, type CAPSLOCK spl CAPSLOCK to insert it.
 
 /**\brief Analyse a token.
    \param str A UTF16-string.
